chore: move mixin extension
