+++ src/main/java/me/aleksilassila/litematica/printer/printer/ActionManager.java
@@ -3,7 +3,7 @@ package me.aleksilassila.litematica.printer.printer;
 import lombok.Setter;
 import me.aleksilassila.litematica.printer.Reference;
 import me.aleksilassila.litematica.printer.config.Configs;
-import me.aleksilassila.litematica.printer.mixin_extension.MultiPlayerGameModeExtension;
+import me.aleksilassila.litematica.printer.mixin.extension.MultiPlayerGameModeExtension;
 import me.aleksilassila.litematica.printer.utils.BlockUtils;
 import me.aleksilassila.litematica.printer.utils.PacketUtils;
 import net.minecraft.client.player.LocalPlayer;
