+++ src/main/resources/litematica-printer.mixins.json
@@ -1,32 +1,32 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "me.aleksilassila.litematica.printer.mixin",
+  "package": "me.aleksilassila.litematica.printer.mixin.printer",
   "compatibilityLevel": "${compatibility_level}",
   "mixins": [
-    "printer.litematica.EasyPlaceUtilsAccessor",
-    "printer.litematica.InventoryUtilsAccessor",
-    "printer.litematica.MixinInventoryUtils",
-    "printer.litematica.MixinMaterialListUtils",
-    "printer.malilib.config.MixinConfigBase",
-    "printer.malilib.config.MixinIConfigBase",
-    "printer.mc.MixinBlockItem",
-    "printer.mc.MixinConnection",
-    "printer.mc.MixinGui",
-    "printer.mc.MixinHud",
-    "printer.mc.ServerboundMovePlayerPacketAccessor"
+    "litematica.EasyPlaceUtilsAccessor",
+    "litematica.InventoryUtilsAccessor",
+    "litematica.MixinInventoryUtils",
+    "litematica.MixinMaterialListUtils",
+    "malilib.config.MixinConfigBase",
+    "malilib.config.MixinIConfigBase",
+    "mc.MixinBlockItem",
+    "mc.MixinConnection",
+    "mc.MixinGui",
+    "mc.MixinHud",
+    "mc.ServerboundMovePlayerPacketAccessor"
   ],
   "client": [
-    "printer.litematica.gui.MixinLitematicaGuiMainMenu",
-    "printer.malilib.MixinAbstractCraftingMenu",
-    "printer.malilib.MixinGuiStringListEdit",
-    "printer.mc.MixinClientCommonPacketListener",
-    "printer.mc.MixinClientLevel",
-    "printer.mc.MixinClientPacketListener",
-    "printer.mc.MixinGui",
-    "printer.mc.MixinLocalPlayer",
-    "printer.mc.MixinMinecraft",
-    "printer.mc.MixinMultiPlayerGameMode"
+    "litematica.gui.MixinLitematicaGuiMainMenu",
+    "malilib.MixinAbstractCraftingMenu",
+    "malilib.MixinGuiStringListEdit",
+    "mc.MixinClientCommonPacketListener",
+    "mc.MixinClientLevel",
+    "mc.MixinClientPacketListener",
+    "mc.MixinGui",
+    "mc.MixinLocalPlayer",
+    "mc.MixinMinecraft",
+    "mc.MixinMultiPlayerGameMode"
   ],
   "injectors": {
     "defaultRequire": 1
